[[8181]] Move a File to Trash or Recycle Bin
Added a sentence to the end of the Warniing + the accompanying example.

NOTE: I've only tested the code for OS X as I do not have a Windows machine. The VB script was extracted from the Use List and dates back to 2008 so verification that it is still valid is required. Sorry, my research for Linux indicated that the Trash location changed with distribution and/or windowing environment so no Linux example is given.

Bug 8181 refers to a Caution for both delete file and delete folder but no such Caution (now Warning) exists in the current documents for delete folder - nor should one exist as you can only delete an empty folder with this command. I therefore see no need to add an example of how to move an empty folder to the Trash or Recycle Bin. I intend to add an example of how to move a 'full' folder to the Trash for the revDeleteFolder command; at which point bug 8181 should be closed.