[[BUG 15176]] Corrected reference to Trash on OS X
BUG 15176 confirms that on OS X the revDeleteFolder is a permenant delete - as per other platforms. Removed the incorrect statement indicating on OS X the folder was trashed.

BUG 8181 requests an example on how to safely move folders to the Trash. Added an example on how to do that. The OS X example has been tested. I do not have a Windows machine, the Windows example was extracted from the Use List and dates back to 2008 so needs verification.

My search on the Use List indicated the Trash on Linux was fluid, dependent of distribution and widnowing system so no Linux example is given.

I have a pull request for 'delete file' which also is in relation to BUG 8181. Once both these pull requests are merged BUG 8181 can also be closed.